Fix result code table mapping.

+++ Engine/Src/SFCore/ResultCode/SFResultCodeSystem.cpp
@@ -10,7 +10,7 @@
 
 #pragma once
 #include "SFTypedefs.h"
-#include "ResultCode/ResultTable.h"
+#include "ResultCode/SFResultTable.h"
 #include "ResultCode/SFResultCodeSystem.h"
 
 
@@ -22,16 +22,16 @@ namespace ResultCode {
 
 		auto& codeMap = ResultTable::GetOrAddFacility(0);
 
-		codeMap.Insert(SUCCESS, {"SUCCESS", "SUCCESS"});
-		codeMap.Insert(SUCCESS_FALSE, {"SUCCESS_FALSE", "SUCCESS_FALSE"});
-		codeMap.Insert(SUCCESS_YIELD, {"SUCCESS_YIELD", "SUCCESS_YIELD"});
-		codeMap.Insert(INVALID_POINTER, {"INVALID_POINTER", "INVALID_POINTER"});
-		codeMap.Insert(FAIL, {"FAIL", "FAIL"});
-		codeMap.Insert(NOT_IMPLEMENTED, {"NOT_IMPLEMENTED", "NOT_IMPLEMENTED"});
-		codeMap.Insert(OUT_OF_MEMORY, {"OUT_OF_MEMORY", "OUT_OF_MEMORY"});
-		codeMap.Insert(UNEXPECTED, {"UNEXPECTED", "UNEXPECTED"});
-		codeMap.Insert(INVALID_ARG, {"INVALID_ARG", "INVALID_ARG"});
-		codeMap.Insert(NO_DATA_EXIST, {"NO_DATA_EXIST", "NO_DATA_EXIST"});
+		codeMap.Insert((uint32_t)SUCCESS, {"SUCCESS", "SUCCESS"});
+		codeMap.Insert((uint32_t)SUCCESS_FALSE, {"SUCCESS_FALSE", "SUCCESS_FALSE"});
+		codeMap.Insert((uint32_t)SUCCESS_YIELD, {"SUCCESS_YIELD", "SUCCESS_YIELD"});
+		codeMap.Insert((uint32_t)INVALID_POINTER, {"INVALID_POINTER", "INVALID_POINTER"});
+		codeMap.Insert((uint32_t)FAIL, {"FAIL", "FAIL"});
+		codeMap.Insert((uint32_t)NOT_IMPLEMENTED, {"NOT_IMPLEMENTED", "NOT_IMPLEMENTED"});
+		codeMap.Insert((uint32_t)OUT_OF_MEMORY, {"OUT_OF_MEMORY", "OUT_OF_MEMORY"});
+		codeMap.Insert((uint32_t)UNEXPECTED, {"UNEXPECTED", "UNEXPECTED"});
+		codeMap.Insert((uint32_t)INVALID_ARG, {"INVALID_ARG", "INVALID_ARG"});
+		codeMap.Insert((uint32_t)NO_DATA_EXIST, {"NO_DATA_EXIST", "NO_DATA_EXIST"});
 
 	}//InitializeResultCode 
 }//namespace ResultCode
